+++ src/main/typescript/components/form/parts/basicInformation/AudienceFieldArray.tsx
@@ -15,7 +15,7 @@
  */
 import * as React from "react"
 import LoadDropdownData from "../../../../lib/formComponents/LoadDropdownData"
-import { FieldArrayProps } from "../../../../lib/formComponents/RepeatableField"
+import { FieldArrayProps } from "../../../../lib/formComponents/ReduxFormUtils"
 import DropdownFieldArray from "../../../../lib/formComponents/DropDownFieldArray"
 import { SchemedValue } from "../../../../lib/metadata/Value"
 import { DropdownList } from "../../../../model/DropdownLists"
